requirements_actualizado.txt
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
asynctest==0.13.0
attrs==24.2.0
backports.zoneinfo==0.2.1
cattrs==23.1.2
certifi==2024.8.30
charset-normalizer==3.4.0
datasets==2.13.2
defusedxml==0.7.1
diff-match-patch==20241021
dill==0.3.6
Django==3.2.25
django-admin==2.0.2
django-excel-response2==3.0.6
django-grappelli==2.15.7
django-import-export==3.2.0
django-six==1.0.5
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
docstring-to-markdown==0.15
et-xmlfile==1.1.0
excel-base==1.0.4
exceptiongroup==1.2.2
filelock==3.12.2
frozenlist==1.3.3
fsspec==2023.1.0
huggingface-hub==0.16.4
idna==3.10
importlib-metadata==3.10.1
isoweek==1.3.3
jedi==0.18.2
jedi-language-server==0.40.0
lsprotocol==2023.0.1
MarkupPy==1.14
multidict==6.0.5
multiprocess==0.70.14
mysqlclient==2.1.1
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
odfpy==1.4.1
packaging==24.0
pandas==1.3.5
parso==0.8.4
Pillow==9.5.0
pyarrow==12.0.1
pydantic==1.10.21
pygls==1.0.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
safetensors==0.4.5
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
tablib==3.4.0
TimeConvert==3.0.13
tokenizers==0.13.3
tqdm==4.67.0
typeguard==2.13.3
typing_extensions==4.7.1
tzlocal==5.1
urllib3==2.0.7
xlrd==2.0.1
xlwt==1.3.0
xxhash==3.5.0
yarl==1.9.4
zipp==3.15.0
